Completed part 4b exercise 4.8: Blog list tests, step1
