+++ part4/blog/tests/test_helper.js
@@ -1,0 +1,70 @@
+const premadeBlogs = [
+    {
+        id: '5a422a851b54a676234d17f7',
+        title: 'React patterns',
+        author: 'Michael Chan',
+        url: 'https://reactpatterns.com/',
+        likes: 7,
+    },
+    {
+        id: '5a422aa71b54a676234d17f8',
+        title: 'Go To Statement Considered Harmful',
+        author: 'Edsger W. Dijkstra',
+        url: 'http://www.u.arizona.edu/~rubinson/copyright_violations/Go_To_Considered_Harmful.html',
+        likes: 5,
+    },
+    {
+        id: '5a422b3a1b54a676234d17f9',
+        title: 'Canonical string reduction',
+        author: 'Edsger W. Dijkstra',
+        url: 'http://www.cs.utexas.edu/~EWD/transcriptions/EWD08xx/EWD808.html',
+        likes: 12,
+    },
+    {
+        id: '5a422b891b54a676234d17fa',
+        title: 'First class tests',
+        author: 'Robert C. Martin',
+        url: 'http://blog.cleancoder.com/uncle-bob/2017/05/05/TestDefinitions.htmll',
+        likes: 10,
+    },
+    {
+        id: '5a422ba71b54a676234d17fb',
+        title: 'TDD harms architecture',
+        author: 'Robert C. Martin',
+        url: 'http://blog.cleancoder.com/uncle-bob/2017/03/03/TDD-Harms-Architecture.html',
+        likes: 0,
+    },
+    {
+        id: '5a422bc61b54a676234d17fc',
+        title: 'Type wars',
+        author: 'Robert C. Martin',
+        url: 'http://blog.cleancoder.com/uncle-bob/2016/05/01/TypeWars.html',
+        likes: 2,
+    }
+]
+
+const oneExtraDijkstraBlog = [
+    {
+        id: '5a422bc61b54a676234d1666',
+        title: 'A fake blog entry',
+        author: 'Edsger W. Dijkstra',
+        url: 'https://www.fakeblog.com',
+        likes: 12
+    }
+]
+
+const oneExtraMartinBlog = [
+    {
+        id: '33422bc61b54a676234d1444',
+        title: 'some more likes',
+        author: 'Robert C. Martin',
+        url: 'https://www.anotherfakeblog.com',
+        likes: 5
+    }
+]
+
+module.exports = {
+    premadeBlogs,
+    oneExtraDijkstraBlog,
+    oneExtraMartinBlog
+}
